refactor(core): share accessibility polling
